Extract base job class and use template methods
This makes the algorithm more obvious and job classes less boilerplate-ish.
